Add calendar/time-mode support & AstroDateTimeEdit
Introduce explicit calendar and time-mode handling across the astrological engine: add CalendarType (Auto/Gregorian/Julian) and TimeMode (ZoneTime/LMT/LAT) to InputData, change timezone to fractional hours (double), and wire these throughout calculations. Add equation-of-time and LAT/LMT helper functions, update Julian/UTC conversion APIs (getJulianDate, getUTfromET, dateTimeFromJulian) to accept calendar type, and adjust usages to preserve historical-date correctness and LAT/LMT conversions. Add a new AstroDateTimeEdit widget for robust date/time entry (including BC years and calendar suffixes) and register it in CMake. Persist calendar/timeMode in AstroFile save/load and update output formatting to annotate calendar/time and display EoT. Also add .github/copilot-instructions.md documentation. These are API-affecting changes — rebuild required; defaults keep backward compatibility (Cal_Auto / Time_ZoneTime).

+++ astroprocessor/src/astro-output.h
@@ -106,7 +106,7 @@ describeSpeculum(const Horoscope&    scope,
                  bool                showFixedStars,
                  SpeculumDisplayMode displayMode = DisplayLocalTime);
 QString
-_formatTime(const QDateTime& dt, short tz);
+_formatTime(const QDateTime& dt, double tz);
 QString
 describe(AstroFileList&& scopes,
          Articles        article  = Article_All,
